Stats overhaul

+++ MicroCy/RegionReporterResult.cs
@@ -7,7 +7,7 @@ namespace DIOS.Core
   [Serializable]
   public class RegionReporterResult
   {
-    public ushort regionNumber;
+    public int regionNumber;
     public List<float> ReporterValues = new List<float>(CAPACITY);
     //public List<float> RP1bgnd = new List<float>(CAPACITY);
     public const int CAPACITY = 40000;
